Ver0.0.3 Optimized starrail configure to use multiprocessing Managers (speedup in local game search). Added feature to downscale screen feature matching threshold based on the native screen resolution. Added logout feature.

+++ src/starrail/honkai_star_rail.py
@@ -56,7 +56,14 @@ class HonkaiStarRail:
         return False
     
     
-    def terminate(self) -> bool:
+    def terminate(self, force_terminate=False) -> bool:
+        if force_terminate:
+            for proc in psutil.process_iter(['pid', 'name']):
+                if proc.info['name'] == GAME_DEFAULT:
+                    proc.terminate()
+                    return True
+            return False
+        
         if self.process.is_running:
             self.process.terminate()
             self.process.wait()
